max piece count is int32

+++ rain/transfer.go
@@ -21,7 +21,7 @@ func newTransfer(tor *TorrentFile, where string) *transfer {
 		torrentFile: tor,
 		where:       where,
 		pieces:      newPieces(tor),
-		bitfield:    NewBitField(nil, int64(tor.NumPieces)),
+		bitfield:    NewBitField(nil, tor.NumPieces),
 	}
 }
 
@@ -79,7 +79,7 @@ func newPieces(tor *TorrentFile) []*piece {
 		if mod != 0 {
 			numBlocks++
 		}
-		p.bitfield = NewBitField(nil, int64(numBlocks))
+		p.bitfield = NewBitField(nil, numBlocks)
 		p.blocks = make([]*block, numBlocks)
 		for j := int32(0); j < div; j++ {
 			p.blocks[j] = &block{
@@ -90,7 +90,7 @@ func newPieces(tor *TorrentFile) []*piece {
 		if mod != 0 {
 			p.blocks[numBlocks-1] = &block{
 				index:  numBlocks - 1,
-				length: mod,
+				length: int32(mod),
 			}
 		}
